+++ test/emu/nosv/flag-enable-all.driver.sh
@@ -1,0 +1,4 @@
+# Test the -a flag with a program that doesn't enable the required models for
+# the events in the stream.
+$OVNI_TEST_BIN
+ovniemu -a ovni